added setuptools and tox testing
setuptools added for packaging and moved requirements
there to be soon able to publish viewcontrol on pypi

pytest added options
- to disable load
- use a custom test directory than return by tmp_path_factory

tox testing covers python 3.6, 3.7 and 3.8

pyinput will only be loaded when a Display is detected,
Keys are saved as string only now. This way viewcontrol
and show can be run on headless systems without causing
errors.